requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
boto3==1.28.0
flask-cors==3.0.10
python-dotenv==1.0.0
tensorflow==2.13.0
numpy==1.23.5
pillow==9.5.0
matplotlib==3.7.1
orjson==3.9.0
python-multipart